@@ -4,7 +4,6 @@ import (
 	"database/sql"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/url"
 	"strconv"
 	"strings"
@@ -153,7 +152,7 @@ func (ch *ClickHouse) Run(r io.Reader) error {
 		return err
 	}
 
-	migration, err := ioutil.ReadAll(r)
+	migration, err := io.ReadAll(r)
 	if err != nil {
 		return err
 	}
