@@ -3,7 +3,6 @@ package file
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
 	"path/filepath"
@@ -13,15 +12,7 @@ import (
 )
 
 func Test(t *testing.T) {
-	tmpDir, err := ioutil.TempDir("", "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		if err := os.RemoveAll(tmpDir); err != nil {
-			t.Error(err)
-		}
-	}()
+	tmpDir := t.TempDir()
 
 	// write files that meet driver test requirements
 	mustWriteFile(t, tmpDir, "1_foobar.up.sql", "1 up")
@@ -47,15 +38,7 @@ func Test(t *testing.T) {
 }
 
 func TestOpen(t *testing.T) {
-	tmpDir, err := ioutil.TempDir("", "TestOpen")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		if err := os.RemoveAll(tmpDir); err != nil {
-			t.Error(err)
-		}
-	}()
+	tmpDir := t.TempDir()
 
 	mustWriteFile(t, tmpDir, "1_foobar.up.sql", "")
 	mustWriteFile(t, tmpDir, "1_foobar.down.sql", "")
@@ -65,22 +48,14 @@ func TestOpen(t *testing.T) {
 	}
 
 	f := &File{}
-	_, err = f.Open("file://" + tmpDir) // absolute path
+	_, err := f.Open("file://" + tmpDir) // absolute path
 	if err != nil {
 		t.Fatal(err)
 	}
 }
 
 func TestOpenWithRelativePath(t *testing.T) {
-	tmpDir, err := ioutil.TempDir("", "TestOpen")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		if err := os.RemoveAll(tmpDir); err != nil {
-			t.Error(err)
-		}
-	}()
+	tmpDir := t.TempDir()
 
 	wd, err := os.Getwd()
 	if err != nil {
@@ -144,36 +119,20 @@ func TestOpenDefaultsToCurrentDirectory(t *testing.T) {
 }
 
 func TestOpenWithDuplicateVersion(t *testing.T) {
-	tmpDir, err := ioutil.TempDir("", "TestOpenWithDuplicateVersion")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		if err := os.RemoveAll(tmpDir); err != nil {
-			t.Error(err)
-		}
-	}()
+	tmpDir := t.TempDir()
 
 	mustWriteFile(t, tmpDir, "1_foo.up.sql", "") // 1 up
 	mustWriteFile(t, tmpDir, "1_bar.up.sql", "") // 1 up
 
 	f := &File{}
-	_, err = f.Open("file://" + tmpDir)
+	_, err := f.Open("file://" + tmpDir)
 	if err == nil {
 		t.Fatal("expected err")
 	}
 }
 
 func TestClose(t *testing.T) {
-	tmpDir, err := ioutil.TempDir("", "TestOpen")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		if err := os.RemoveAll(tmpDir); err != nil {
-			t.Error(err)
-		}
-	}()
+	tmpDir := t.TempDir()
 
 	f := &File{}
 	d, err := f.Open("file://" + tmpDir)
@@ -187,16 +146,13 @@ func TestClose(t *testing.T) {
 }
 
 func mustWriteFile(t testing.TB, dir, file string, body string) {
-	if err := ioutil.WriteFile(path.Join(dir, file), []byte(body), 06444); err != nil {
+	if err := os.WriteFile(path.Join(dir, file), []byte(body), 06444); err != nil {
 		t.Fatal(err)
 	}
 }
 
 func mustCreateBenchmarkDir(t *testing.B) (dir string) {
-	tmpDir, err := ioutil.TempDir("", "Benchmark")
-	if err != nil {
-		t.Fatal(err)
-	}
+	tmpDir := t.TempDir()
 
 	for i := 0; i < 1000; i++ {
 		mustWriteFile(t, tmpDir, fmt.Sprintf("%v_foobar.up.sql", i), "")
