@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 
 	"github.com/golang-migrate/migrate/v4/source"
@@ -102,7 +101,7 @@ func (b *Bindata) ReadUp(version uint) (r io.ReadCloser, identifier string, err 
 		if err != nil {
 			return nil, "", err
 		}
-		return ioutil.NopCloser(bytes.NewReader(body)), m.Identifier, nil
+		return io.NopCloser(bytes.NewReader(body)), m.Identifier, nil
 	}
 	return nil, "", &os.PathError{Op: fmt.Sprintf("read version %v", version), Path: b.path, Err: os.ErrNotExist}
 }
@@ -113,7 +112,7 @@ func (b *Bindata) ReadDown(version uint) (r io.ReadCloser, identifier string, er
 		if err != nil {
 			return nil, "", err
 		}
-		return ioutil.NopCloser(bytes.NewReader(body)), m.Identifier, nil
+		return io.NopCloser(bytes.NewReader(body)), m.Identifier, nil
 	}
 	return nil, "", &os.PathError{Op: fmt.Sprintf("read version %v", version), Path: b.path, Err: os.ErrNotExist}
 }
