@@ -3,8 +3,6 @@ package sqlcipher
 import (
 	"database/sql"
 	"fmt"
-	"io/ioutil"
-	"os"
 	"path/filepath"
 	"testing"
 
@@ -17,15 +15,7 @@ import (
 )
 
 func Test(t *testing.T) {
-	dir, err := ioutil.TempDir("", "sqlite3-driver-test")
-	if err != nil {
-		return
-	}
-	defer func() {
-		if err := os.RemoveAll(dir); err != nil {
-			t.Error(err)
-		}
-	}()
+	dir := t.TempDir()
 	t.Logf("DB path : %s\n", filepath.Join(dir, "sqlite3.db"))
 	p := &Sqlite{}
 	addr := fmt.Sprintf("sqlite3://%s", filepath.Join(dir, "sqlite3.db"))
@@ -37,15 +27,7 @@ func Test(t *testing.T) {
 }
 
 func TestMigrate(t *testing.T) {
-	dir, err := ioutil.TempDir("", "sqlite3-driver-test")
-	if err != nil {
-		return
-	}
-	defer func() {
-		if err := os.RemoveAll(dir); err != nil {
-			t.Error(err)
-		}
-	}()
+	dir := t.TempDir()
 	t.Logf("DB path : %s\n", filepath.Join(dir, "sqlite3.db"))
 
 	db, err := sql.Open("sqlite3", filepath.Join(dir, "sqlite3.db"))
@@ -72,15 +54,7 @@ func TestMigrate(t *testing.T) {
 }
 
 func TestMigrationTable(t *testing.T) {
-	dir, err := ioutil.TempDir("", "sqlite3-driver-test-migration-table")
-	if err != nil {
-		return
-	}
-	defer func() {
-		if err := os.RemoveAll(dir); err != nil {
-			t.Error(err)
-		}
-	}()
+	dir := t.TempDir()
 
 	t.Logf("DB path : %s\n", filepath.Join(dir, "sqlite3.db"))
 
@@ -120,15 +94,7 @@ func TestMigrationTable(t *testing.T) {
 }
 
 func TestNoTxWrap(t *testing.T) {
-	dir, err := ioutil.TempDir("", "sqlite3-driver-test")
-	if err != nil {
-		return
-	}
-	defer func() {
-		if err := os.RemoveAll(dir); err != nil {
-			t.Error(err)
-		}
-	}()
+	dir := t.TempDir()
 	t.Logf("DB path : %s\n", filepath.Join(dir, "sqlite3.db"))
 	p := &Sqlite{}
 	addr := fmt.Sprintf("sqlite3://%s?x-no-tx-wrap=true", filepath.Join(dir, "sqlite3.db"))
@@ -142,19 +108,11 @@ func TestNoTxWrap(t *testing.T) {
 }
 
 func TestNoTxWrapInvalidValue(t *testing.T) {
-	dir, err := ioutil.TempDir("", "sqlite3-driver-test")
-	if err != nil {
-		return
-	}
-	defer func() {
-		if err := os.RemoveAll(dir); err != nil {
-			t.Error(err)
-		}
-	}()
+	dir := t.TempDir()
 	t.Logf("DB path : %s\n", filepath.Join(dir, "sqlite3.db"))
 	p := &Sqlite{}
 	addr := fmt.Sprintf("sqlite3://%s?x-no-tx-wrap=yeppers", filepath.Join(dir, "sqlite3.db"))
-	_, err = p.Open(addr)
+	_, err := p.Open(addr)
 	if assert.Error(t, err) {
 		assert.Contains(t, err.Error(), "x-no-tx-wrap")
 		assert.Contains(t, err.Error(), "invalid syntax")
